Move recipe view to own file
